family-studies - Improve support for family studies Separated Pedigree to interface (Pedigree) and implementation (DefaultPedigree).
